Create VBA Chart Series Update
Currently I have an Excel Workbook that is archiving daily crude oil prices. One of the worksheets is displaying the archived crude oil prices on various charts. Each day a new row is added to the archived list of prices. I would like to make sure the charts each show the most recent day's pricing, 10 day's ago, and 20 days ago.  This code ensures that the correct days' pricing is automatically  included in the charts.